LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-form-attribute.tentative-expected.txt




PASS Tests for parser inserted controls
FAIL [SELECTMENU] Basic form association - control with no form attribute is associated with ancestor assert_equals: expected (object) Element node <form id="form1"><selectmenu></selectmenu></form> but got (undefined) undefined
FAIL [SELECTMENU] Form owner is reset to null when control's form attribute is set to an ID that does not exist in the document assert_equals: expected (object) Element node <form id="form1"><selectmenu form="form1"></selectmenu></... but got (undefined) undefined
FAIL [SELECTMENU] Control whose form attribute is an empty string has no form owner assert_equals: expected (object) null but got (undefined) undefined
FAIL [SELECTMENU] Control whose form attribute is an empty string has no form owner even when form with empty attribute is present assert_equals: expected (object) null but got (undefined) undefined
FAIL [SELECTMENU] Control's form attribute must be a case sensitive match for the form's id assert_equals: expected (object) null but got (undefined) undefined
FAIL [SELECTMENU] Setting the form attribute of a control to the id of a non-ancestor form works assert_equals: expected (object) Element node <form id="form1"><selectmenu></selectmenu></form> but got (undefined) undefined
FAIL [SELECTMENU] Removing form id from a control resets the form owner to ancestor assert_equals: expected (object) Element node <form id="form1"></form> but got (undefined) undefined
FAIL [SELECTMENU] Removing the form owner of a control with form attribute resets the form owner to null assert_equals: expected (object) Element node <form id="form1"></form> but got (undefined) undefined
FAIL [SELECTMENU] Changing form attibute of control resets form owner to correct form assert_equals: expected (object) Element node <form id="form3"><selectmenu></selectmenu></form> but got (undefined) undefined
FAIL [SELECTMENU] Moving a control with form attribute within the document does not change the form owner assert_equals: expected (object) Element node <form id="form1"></form> but got (undefined) undefined
FAIL [SELECTMENU] When the id of a non-ancestor form changes from not being a match for the form attribute to being a match, the control's form owner is reset assert_equals: expected (object) null but got (undefined) undefined
FAIL [SELECTMENU] When form element with same ID as the control's form attribute is inserted earlier in tree order, the form owner is changed to the inserted form assert_equals: expected (object) Element node <form id="form1"><selectmenu form="form1"></selectmenu></... but got (undefined) undefined
FAIL [SELECTMENU] When non-form element with same ID as the control's form attribute is inserted earlier in tree order, the control does not have a form owner assert_equals: expected (object) Element node <form id="form1"></form> but got (undefined) undefined
FAIL [SELECTMENU] A control that is not in the document but has the form attribute set is associated with the nearest ancestor form if one exists assert_equals: expected (object) Element node <form id="form1"><selectmenu form="form1"></selectmenu></... but got (undefined) undefined
